- 实现了超时重试，重试逻辑主要在 remoteCall中实现

+++ rpc-client/src/main/java/com/ws/rpc/client/dto/RpcRequestMetaData.java
@@ -20,5 +20,5 @@ public class RpcRequestMetaData {
     private RpcMessage rpcMessage;
     private String serviceAddress;
     private Integer servicePort;
-    private Long timeout;
+    private int timeout;
 }
